Bring back `teleport/DesktopSession`
